+++ src/Graphics/GRRender.h
@@ -140,7 +140,7 @@ public:
 	virtual void SetVertexFormat(const GRVertexElement* e){}
 	//virtual void SetVertexShader(void* shader){}
 	virtual void DrawDirect(GRRenderBaseIf::TPrimitiveType ty, void* vtx, size_t count, size_t stride=0){}
-	virtual void DrawIndexed(GRRenderBaseIf::TPrimitiveType ty, size_t* idx, void* vtx, size_t count, size_t stride=0){}
+	virtual void DrawIndexed(GRRenderBaseIf::TPrimitiveType ty, GLuint* idx, void* vtx, size_t count, size_t stride=0){}
 	virtual void DrawArrays(GRRenderBaseIf::TPrimitiveType ty, GRVertexArray* arrays, size_t count){}
  	virtual void DrawArrays(GRRenderBaseIf::TPrimitiveType ty, size_t* idx, GRVertexArray* arrays, size_t count){}
 	virtual void DrawPoint(Vec3f p){}
@@ -234,7 +234,7 @@ public:
 	virtual void DrawDirect(GRRenderBaseIf::TPrimitiveType ty, void* vtx, size_t ct, size_t st=0)			\
 		{ ptr DrawDirect(ty, vtx, ct, st); }																\
 	virtual void DrawIndexed(GRRenderBaseIf::TPrimitiveType ty,												\
-		size_t* idx, void* vtx, size_t ct, size_t st=0)														\
+		GLuint* idx, void* vtx, size_t ct, size_t st=0)														\
 		{ ptr DrawIndexed(ty, idx, vtx, ct, st); }															\
 	virtual void DrawPoint(Vec3f p){ ptr DrawPoint(p); }													\
 	virtual void DrawLine(Vec3f p0, Vec3f p1){ ptr DrawLine(p0, p1); }										\
