+++ src/GraphicsD3D/GRDeviceD3D.h
@@ -96,7 +96,7 @@ public:
 	///	頂点座標を指定してプリミティブを描画
 	virtual void DrawDirect(GRRenderBaseIf::TPrimitiveType ty, void* begin, size_t count, size_t stride=0);
 	///	頂点座標とインデックスを指定してプリミティブを描画
-	virtual void DrawIndexed(GRRenderBaseIf::TPrimitiveType ty, size_t* idx, void* vtx, size_t count, size_t stride=0);
+	virtual void DrawIndexed(GRRenderBaseIf::TPrimitiveType ty, GLuint* idx, void* vtx, size_t count, size_t stride=0);
 	///	DisplayListの表示
 	virtual void DrawList(int i);
 	///	DisplayListの解放
